requirements.txt
alembic==1.7.4
certifi==2021.5.30
chardet==4.0.0
greenlet==1.1.2
idna==2.10
Mako==1.1.5
MarkupSafe==2.0.1
oauthlib==3.1.1
peewee==3.14.4
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.26
tornado==6.1
urllib3==1.26.5
